Add files via upload
This is the final stage of the program. I included a way to print the space group of the perturbed and unperturbed cell while also allowing the user to define which phonons he wants calculated. I clean up much of the code and made it more maintainable. Finally, if the user does not specify what phonons they want calculated, all phonons will be calculated